Improvements to build infrastructure. Now it is possible to --purge before build, and building all folders plus individual folder builds are possible.
Build all folders after a purge of _build:
./bin/build.sh --purge

Build only folder 02 without purge:
./bin/build.sh 02

Build only folder 01 but purge _build first:
./bin/build.sh 02

+++ 2024/bin/build.sh
@@ -1,21 +1,39 @@
 #!/bin/bash
 
-rm -rf _build
+export AOC_DATA_PATH="/home/dad/src/aoc/aoc-data/2024/"
+
+purge=false
+
+if [ "$1" == "--purge" ]; then
+	purge=true
+	shift
+fi
+
+folder=$1
+
+if $purge; then
+	rm -rf _build
+fi
 
 build_folder() {
 	local folder=$1
 	mkdir -p _build/$folder
 	cd _build/$folder
-	cmake ../../
+	echo "Building in folder: $(pwd)"
+	cmake -DTARGET_DIR=$folder ../..
 	make
 	cd ../..
 }
 
-if [ -z "$1" ]; then
-	for folder in $(ls -d */); do
-		folder_name=$(basename $folder)
-		build_folder $folder_name
+echo "Current path: $(pwd)"
+echo "AOC_DATA_PATH: $AOC_DATA_PATH"
+
+if [ -z "$folder" ]; then
+	for dir in [0-9]*; do
+		if [ -d "$dir" ]; then
+			build_folder "$dir"
+		fi
 	done
 else
-	build_folder $1
+	build_folder "$folder"
 fi
